backend/requirements.txt
Flask==2.0.1
redis==3.5.3
psycopg2-binary==2.9.1
flask-lambda==0.0.4